+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
   - Update script
 
 ### Changed ###
+- **Do not terminate the DUC on**:
+  - Unknown server response
 - **The config file interpreter**
 
 ### Fixed ###